refactor TopicProperties
